+++ cx-portal/src/features/notification/apiSlice.ts
@@ -1,24 +1,28 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import UserService from 'services/UserService'
+import { apiBaseQuery } from 'utils/rtkUtil'
 import { CXNotification } from './types'
 
 export const apiSlice = createApi({
   reducerPath: 'info/notifications',
-  baseQuery: fetchBaseQuery({
-    baseUrl: 'https://portal.dev.demo.catena-x.net/assets/api/notification', //getApiBase(),
-    prepareHeaders: (headers) => {
-      headers.set('authorization', `Bearer ${UserService.getToken()}`)
-      return headers
-    },
-  }),
+  baseQuery: fetchBaseQuery(apiBaseQuery()),
   endpoints: (builder) => ({
-    getCount: builder.query<number, null>({
-      query: () => '/count.json',
+    getNotificationCount: builder.query<number, boolean>({
+      query: (read) => `/api/notification/count?read=${read}`,
     }),
     getNotifications: builder.query<CXNotification[], null>({
-      query: () => '/all.json',
+      query: () => '/api/notification',
+    }),
+    setNotificationRead: builder.mutation<void, string>({
+      query: (id) => ({
+        url: `/api/notification/${id}/read`,
+        method: 'PUT',
+      }),
     }),
   }),
 })
 
-export const { useGetCountQuery, useGetNotificationsQuery } = apiSlice
+export const {
+  useGetNotificationCountQuery,
+  useGetNotificationsQuery,
+  useSetNotificationReadMutation,
+} = apiSlice